Relocate App component. Replace with original WebJive one
