--- src/App.css
@@ -1,35 +1,0 @@
-.App {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-
-  display: grid;
-  grid-template-rows: 4em auto;
-  grid-template-columns: 30em auto;
-}
-
-.App header {
-  background-color: #222;
-  padding: 20px;
-  color: white;
-  text-align: center;
-
-  grid-row: 1;
-  grid-column: 1 / -1;
-}
-
-.App .left-column {
-  grid-row: 2;
-  grid-column: 1;
-  overflow-y: auto;
-}
-
-.App .right-column {
-  grid-row: 2;
-  grid-column: 2;
-}
-
-.App .double-column {
-  grid-row: 2;
-  grid-column: 1 / -1;
-}
