+++ src/components/App/App.css
@@ -1,32 +1,35 @@
 .App {
-  text-align: center;
+  position: absolute;
+  height: 100%;
+  width: 100%;
+
+  display: grid;
+  grid-template-rows: 4em auto;
+  grid-template-columns: 30em auto;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
+.App header {
+  background-color: #222;
+  padding: 20px;
   color: white;
+  text-align: center;
+
+  grid-row: 1;
+  grid-column: 1 / -1;
 }
 
-.App-link {
-  color: #61dafb;
+.App .left-column {
+  grid-row: 2;
+  grid-column: 1;
+  overflow-y: auto;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.App .right-column {
+  grid-row: 2;
+  grid-column: 2;
+}
+
+.App .double-column {
+  grid-row: 2;
+  grid-column: 1 / -1;
 }
