--- src/components/App/App.jsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import logo from '../../resources/logo.svg';
-import './App.css';
-
-const App = () => (
-  <div className="App">
-    <header className="App-header">
-      <img src={logo} className="App-logo" alt="logo" />
-      <p>
-        Edit&nbsp;
-        <code>src/App.js</code>
-        &nbsp;and save to reload.
-      </p>
-      <a className="App-link" href="https://reactjs.org" target="_blank" rel="noopener noreferrer">
-        Learn React
-      </a>
-    </header>
-  </div>
-);
-
-export default App;
