+++ src/index.tsx
@@ -3,15 +3,15 @@ import * as React from 'react';
 import { render } from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import configureStore from './store/configureStore';
-import {Provider} from 'react-redux';
-import App from './App';
+import { Provider } from 'react-redux';
+import App from './components/App/App';
 import './index.css';
 
 const store = configureStore();
 
 render(
   <Provider store={store}>
-      <App />
+    <App />
   </Provider>,
   document.getElementById('root')
 );
